src/main/python/cn/gov/forestry/requirements.txt
requests>=2.28.2

# Scheduler
apscheduler>=3.10.1

# Web server
Flask>=2.3.2

# Excel processing
openpyxl>=3.1.2

# Shapefile and GIS
pyshp>=2.3.0
shapely>=2.0.1
rasterio>=1.3.8
affine>=2.4.0

# Optional / Development
pytest>=7.4.0
python-dotenv>=1.0.0

